pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "streamlit>=1.45.0",
    "tensorflow>=2.14.0",
]
